+++ modules/common/guest-platform/src/platform/builtins/kotlin/PrimitiveTypes.kt
@@ -23,15 +23,27 @@ public class Char private constructor() : Comparable<Char> {
 public class Int private constructor() : Number(), Comparable<Int> {
     public external override operator fun compareTo(other: Int): Int
 
+    public external operator fun compareTo(other: Long): Int
+
     public external operator fun plus(other: Int): Int
+
+    public external operator fun plus(other: Long): Long
 
     public external operator fun minus(other: Int): Int
 
+    public external operator fun minus(other: Long): Long
+
     public external operator fun times(other: Int): Int
+
+    public external operator fun times(other: Long): Long
 
     public external operator fun div(other: Int): Int
 
+    public external operator fun div(other: Long): Long
+
     public external operator fun rem(other: Int): Int
+
+    public external operator fun rem(other: Long): Long
 
     public external operator fun unaryMinus(): Int
 
@@ -51,6 +63,8 @@ public class Int private constructor() : Number(), Comparable<Int> {
 
     public external fun toChar(): Char
 
+    public external fun toLong(): Long
+
     public companion object {
         public const val MIN_VALUE: Int = -2147483647 - 1
         public const val MAX_VALUE: Int = 2147483647
@@ -61,7 +75,54 @@ internal class Byte private constructor() : Number()
 
 internal class Short private constructor() : Number()
 
-internal class Long private constructor() : Number()
+public class Long private constructor() : Number(), Comparable<Long> {
+    public external override operator fun compareTo(other: Long): Int
+
+    public external operator fun compareTo(other: Int): Int
+
+    public external operator fun plus(other: Long): Long
+
+    public external operator fun plus(other: Int): Long
+
+    public external operator fun minus(other: Long): Long
+
+    public external operator fun minus(other: Int): Long
+
+    public external operator fun times(other: Long): Long
+
+    public external operator fun times(other: Int): Long
+
+    public external operator fun div(other: Long): Long
+
+    public external operator fun div(other: Int): Long
+
+    public external operator fun rem(other: Long): Long
+
+    public external operator fun rem(other: Int): Long
+
+    public external operator fun unaryMinus(): Long
+
+    public external infix fun and(other: Long): Long
+
+    public external infix fun or(other: Long): Long
+
+    public external infix fun xor(other: Long): Long
+
+    public external fun inv(): Long
+
+    public external infix fun shl(bitCount: Int): Long
+
+    public external infix fun shr(bitCount: Int): Long
+
+    public external infix fun ushr(bitCount: Int): Long
+
+    public external fun toInt(): Int
+
+    public companion object {
+        public const val MIN_VALUE: Long = -9223372036854775807L - 1L
+        public const val MAX_VALUE: Long = 9223372036854775807L
+    }
+}
 
 internal class Float private constructor() : Number()