+++ modules/common/guest-platform/src/platform/libraries/std-terminal/compukter/io/StdioSupport.kt
@@ -14,6 +14,8 @@ internal fun stdoutPrint(value: String): Unit = StdioBindings.write(value)
 
 internal fun stdoutPrint(value: Int): Unit = StdioBindings.write(stdoutInt(value))
 
+internal fun stdoutPrint(value: Long): Unit = StdioBindings.write("$value")
+
 internal fun stdoutPrint(value: Boolean): Unit = StdioBindings.write(if (value) "true" else "false")
 
 internal fun stdoutPrint(value: Char): Unit = StdioBindings.write(stdoutChar(value))
@@ -23,6 +25,8 @@ internal fun stdoutPrintln(): Unit = StdioBindings.write("\n")
 internal fun stdoutPrintln(value: String): Unit = StdioBindings.write(value + "\n")
 
 internal fun stdoutPrintln(value: Int): Unit = StdioBindings.write(stdoutInt(value) + "\n")
+
+internal fun stdoutPrintln(value: Long): Unit = StdioBindings.write("$value\n")
 
 internal fun stdoutPrintln(value: Boolean): Unit = StdioBindings.write(if (value) "true\n" else "false\n")